Moved retry/nonretry errors to general errors package.
