+++ runtime/errors/non_retry_error.go
@@ -4,15 +4,6 @@ import (
 	"fmt"
 )
 
-// Default errors for the error handler.
-type RetryError struct {
-	Message string
-}
-
-func (e *RetryError) Error() string {
-	return fmt.Sprintf("RetryError: %s", e.Message)
-}
-
 type NonRetryError struct {
 	Message string
 }
