+++ runtime/errors/retry_error.go
@@ -1,0 +1,13 @@
+package errors
+
+import (
+	"fmt"
+)
+
+type RetryError struct {
+	Message string
+}
+
+func (e *RetryError) Error() string {
+	return fmt.Sprintf("RetryError: %s", e.Message)
+}
